Make network communication dependencies optional.
This removal was supposed to be part of a past commit
(cd1d0cb137b74c6baec397c703f4d1c8830427c4) but had in
fact not been properly finished.

With this commit, install declearn no longer results
in the automatic installation of gRPC and websockets,
which are properly relegated to being optional, user-
triggered-installation, dependencies.