Prune 'test_main.py' pytest-collected test scenarios.
The legacy 'test_main.py' integration tests are now mostly redundant
with other integration tests, adding little marginal value to the CI/CD
tests pipeline. From a Coverage POV, the only thing that is not already
tested elsewhere is the actual use of Scaffold.

Hence the test scenarios collected during PyTest runs are restricted to:
- FedAvg and Scaffold variants of Optimization strategies.
  (prior: FedAvg + FedAvgM + Scaffold)
- In non-fulltest, only the Scaffold strategy is used.
  (prior: FedAvg + Scaffold)

This results in testing 3 scenarios in the limited pipeline and 18 in the
"fulltest" one, when there previously would be 6 or 27 scenarios.