Deprecate 'axis' and 'keepdims' kwargs of 'Vector.sum'.
These arguments have never been used, are defined at the level of
subclasses rather than that of the 'Vector' ABC, and are bound to
create some issues if mis-used.

They are being deprecated, and will be considered for re-instatement
with proper ABC definition and documentation if and only if they are
required to implement some given algorithm in the future.