Remove 'Generic' type-hints from SecAgg test suites for pytest.
In Python 3.8 (maybe also in Python 3.9? this was not tested yet),
pytest cannot collect test classes that have a '__new__' method,
which is the case when inheriting from the 'Generic' type-hint.

This commit therefore drops some advances type hints that are of
interest but not critical in making the code understandable, so
that compatibility with our CI runner is restored and tests can
be properly run.