Fix quantization issues based on float values casting.
In numpy 2.0, summing floatN values with an OOB value will result
in inf values, hence making quantization faulty. This was caught
in our unit tests for float16 inputs.
To fix this, the present commit adds backend checks to determine
the minimal float precision required for quantization operations
and use either that or the input float precision.